voice-agent-py/pyproject.toml
[project]
name = "voice-agent-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[deepgram,google,openai,silero]>=1.2.14",
    "python-dotenv>=1.1.1",
]
